pixel_back/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
redis>=4.2.0
asyncpg>=0.25.0
websockets>=10.0
python-dotenv>=0.19.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
pillow~=9.4.0
numpy~=2.2.5